+++ CHANGES.md
@@ -30,6 +30,7 @@ Changes:
 - Add lower court extraction to `federal_appellate` scrapers #1560
 
 - Retrieve lower court information in `delaware` scrapers #1569
+- Retrieve lower court information in `idaho` #1569
 
 Fixes:
 - Fix connappct #1580